requirements.txt
# License Plate Recognition System - Core Dependencies
# Common dependencies shared across all components

# Core Python dependencies
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.7.2

# Machine Learning (Core)
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.232

# Image processing
imageio==2.31.5
scikit-image==0.21.0

# Utilities
PyYAML==6.0.1
tqdm==4.66.1
requests==2.31.0
psutil==5.9.6

# Development tools
pytest==7.4.3
black==23.9.1
flake8==6.1.0

# Configuration management
python-dotenv==1.0.0

# Logging and monitoring
colorlog==6.7.0